refactor: renames files/components
