+++ src/features/user/MFV2/ForestProgress/microComponents/PlantTreeBar.tsx
@@ -1,36 +1,34 @@
-import ConservedAreaTargetIcon from '../../../../../../public/assets/images/icons/Mfv2/ConservAreaTargetIcon';
-import targetBarStyle from '../TreeTargetBar.module.scss';
 import EditTargetIcon from '../../../../../../public/assets/images/icons/Mfv2/EditTargetIcon';
+import TreesPlantedIcon from '../../../../../../public/assets/images/icons/Mfv2/TreesPlantedIcon';
+import targetBarStyle from '../TreeTargetBar.module.scss';
+import themeProperties from '../../../../../theme/themeProperties';
 import { useMyForestV2 } from '../../../../common/Layout/MyForestContextV2';
 import { useTranslations } from 'next-intl';
-import { calculatePercentage } from '../../../../../utils/myForestV2Utils';
+import { getAchievedTarget } from '../../../../../utils/myForestV2Utils';
 import { useMemo } from 'react';
-import themeProperties from '../../../../../theme/themeProperties';
 import { useRouter } from 'next/router';
 
-interface ConservTargetBarProps {
+interface EditButtonProps {
+  handleOpen: () => void;
+}
+export interface TargetBarProps {
   giftPercentage: number;
   personalPercentage: number;
   giftsReceivedCount: number | undefined;
 }
 
-interface EditButtonProps {
-  handleOpen: () => void;
-}
-
 const EditButton = ({ handleOpen }: EditButtonProps) => {
-  const { conservTarget } = useMyForestV2();
+  const { treeTarget } = useMyForestV2();
   const tProfile = useTranslations('Profile');
-  const { mediumBlue } = themeProperties;
   return (
     <div className={targetBarStyle.editTargetButtonContainer}>
       <button
         className={targetBarStyle.editTargetContainer}
         onClick={handleOpen}
       >
-        <EditTargetIcon width={9} color={`${mediumBlue}`} />
-        <p className={targetBarStyle.conservTargetLabel}>
-          {conservTarget > 0
+        <EditTargetIcon width={9} color={themeProperties.primaryDarkColor} />
+        <p className={targetBarStyle.treeTargetLabel}>
+          {treeTarget > 0
             ? tProfile('progressBar.editTarget')
             : tProfile('progressBar.setTarget')}
         </p>
@@ -39,65 +37,64 @@ const EditButton = ({ handleOpen }: EditButtonProps) => {
   );
 };
 
-const ConservTargetBar = ({
+const TreeTargetBar = ({
   giftPercentage,
   personalPercentage,
   giftsReceivedCount,
-}: ConservTargetBarProps) => {
-  const tProfile = useTranslations('Profile');
-  const { conservTarget, conservArea, conservChecked, contributionsResult } =
+}: TargetBarProps) => {
+  const { treePlanted, treeTarget, treeChecked, contributionsResult } =
     useMyForestV2();
-  const giftReceived =
-    contributionsResult?.stats.areaConservedInM2.received ?? 0;
+  const tProfile = useTranslations('Profile');
+  const giftReceived = contributionsResult?.stats.treesDonated.received ?? 0;
 
   return (
     <div className={targetBarStyle.targetSubContainer}>
       <div className={targetBarStyle.statisticsContainer}>
-        <div className={targetBarStyle.iconContainerConservArea}>
-          <ConservedAreaTargetIcon width={19} />
+        <div className={targetBarStyle.iconContainerTreeTarget}>
+          <TreesPlantedIcon width={19} />
         </div>
         <div className={targetBarStyle.targetStatisticsContainer}>
           <div className={targetBarStyle.stat}>
-            {conservTarget > 0 && conservChecked
-              ? tProfile('progressBar.conservWithTarget', {
+            {treeTarget > 0 && treeChecked
+              ? tProfile('progressBar.treeWithTarget', {
                   count:
-                    conservArea !== Math.floor(conservArea)
-                      ? conservArea.toFixed(1)
-                      : conservArea,
-                  unit:
-                    conservTarget !== Math.floor(conservTarget)
-                      ? conservTarget.toFixed(1)
-                      : conservTarget,
+                    treePlanted !== Math.floor(treePlanted)
+                      ? treePlanted.toFixed(1)
+                      : treePlanted,
+                  total:
+                    treeTarget !== Math.floor(treeTarget)
+                      ? treeTarget.toFixed(1)
+                      : treeTarget,
                 })
-              : tProfile('progressBar.conservWithoutTarget', {
-                  unit:
-                    conservArea !== Math.floor(conservArea)
-                      ? conservArea.toFixed(1)
-                      : conservArea,
+              : tProfile('progressBar.treeWithoutTarget', {
+                  count:
+                    treePlanted !== Math.floor(treePlanted)
+                      ? treePlanted.toFixed(1)
+                      : treePlanted,
                 })}
           </div>
           <div className={targetBarStyle.barContainer}>
-            <div className={targetBarStyle.barSubContainerRestoreArea}>
+            <div className={targetBarStyle.barSubContainerTreeTarget}>
               <div
                 style={{
                   width: `${personalPercentage}%`,
                   borderTopRightRadius: `${
-                    giftPercentage !== 0 || conservTarget > conservArea ? 0 : 5
+                    giftPercentage !== 0 || treeTarget > treePlanted ? 0 : 5
                   }px`,
                   borderBottomRightRadius: `${
-                    giftPercentage !== 0 || conservTarget > conservArea ? 0 : 5
+                    giftPercentage !== 0 || treeTarget > treePlanted ? 0 : 5
                   }px`,
                 }}
-                className={targetBarStyle.conservTargetCompleteBar}
+                className={targetBarStyle.treeTargetCompleteBar}
               ></div>
               <div
                 style={{
                   width: `${giftPercentage}%`,
                   borderTopRightRadius: `${
-                    conservTarget > giftReceived ? 0 : 5
+                    treeTarget > giftReceived ? 0 : 5
                   }px`,
                   borderBottomRightRadius: `${
-                    conservTarget > giftReceived ? 0 : 5
+                    treeTarget > giftReceived ? 0 : 5
                   }px`,
                   borderTopLeftRadius: `${personalPercentage === 0 ? 5 : 0}px`,
                   borderBottomLeftRadius: `${
@@ -108,9 +105,7 @@ const ConservTargetBar = ({
               ></div>
             </div>
             <div>
-              {conservTarget > 0 &&
-                conservChecked &&
-                `${giftPercentage + personalPercentage}%`}
+              {treeTarget > 0 && `${giftPercentage + personalPercentage}%`}
             </div>
           </div>
           {giftsReceivedCount !== undefined && giftsReceivedCount > 0 && (
@@ -126,26 +121,23 @@ const ConservTargetBar = ({
   );
 };
 
-interface ConservAreaTargetProp {
-  handleOpen: () => void;
-}
-
-const ConservAreaTarget = ({ handleOpen }: ConservAreaTargetProp) => {
-  const { contributionsResult, conservTarget } = useMyForestV2();
+const PlantTreeBar = ({ handleOpen }: EditButtonProps) => {
   const { asPath } = useRouter();
-  const giftsReceivedCount =
-    contributionsResult?.stats.areaConservedInM2.received;
-  const personal = contributionsResult?.stats.areaConservedInM2.personal;
-  const _calculatePercentage = useMemo(
-    () => calculatePercentage(conservTarget, giftsReceivedCount, personal),
-    [conservTarget, contributionsResult]
-  );
 
+  const { contributionsResult, treeTarget } = useMyForestV2();
+  const giftsReceivedCount =
+    contributionsResult?.stats.treesDonated.received ?? 0;
+  const personal = contributionsResult?.stats.treesDonated.personal ?? 0;
+
+  const _calculatePercentage = useMemo(
+    () => getAchievedTarget(treeTarget, giftsReceivedCount, personal),
+    [treeTarget, contributionsResult]
+  );
   return (
-    <div className={targetBarStyle.targetMainContainerConservArea}>
+    <div className={targetBarStyle.targetMainContainerTreeTarget}>
       {asPath === '/en/profile/mfv2' && <EditButton handleOpen={handleOpen} />}
 
-      <ConservTargetBar
+      <TreeTargetBar
         giftPercentage={Number(_calculatePercentage.giftPercentage.toFixed(1))}
         personalPercentage={Number(
           _calculatePercentage.personalPercentage.toFixed(1)
@@ -156,4 +148,4 @@ const ConservAreaTarget = ({ handleOpen }: ConservAreaTargetProp) => {
   );
 };
 
-export default ConservAreaTarget;
+export default PlantTreeBar;
